\name{40_as_methods}
\alias{as.list.dset}
\title{Convert Distribution Sets to Lists}
\description{
Covert dset objects to lists.\cr
(Refer to \link{Distribution Sets}).

You should wrap a dset object inside the function.\cr
(An exception is plotting).
}
\usage{
\method{as.list}{dset}(x, \dots)
}
\arguments{
\item{x}{A dset (distribution set) object.}
\item{\dots}{Ignored.}
}
\value{A standard list.}
\references{Refer to the vignette for an overview, references and better examples.}
